examples/emr-eks-app-python/requirements.txt
aws-cdk-lib==2.51.0
constructs>=10.0.0,<11.0.0
aws_analytics_reference_architecture==2.9.2
aws-cdk.lambda-layer-kubectl-v22==2.0.7